% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cibap}
\alias{cibap}
\title{CIBAP Report}
\format{
\subsection{\code{cibap}}{

A \link{PdfDocument} with 73 pages; see \link{burgerschap} for the word columns.
}
}
\source{
\url{https://www.rijksoverheid.nl/documenten/rapporten/2024/09/16/kwaliteitsagenda-2024-2027-cibap}
}
\usage{
cibap
}
\description{
The report \verb{Kwaliteitsagenda 2024-2027 Cibap}, published on the
Rijksoverheid website on September 16, 2024, read with \code{\link[=pdf_read]{pdf_read()}}. It
was chosen because it contains several pages with columns, infographics
and different layouts.
}
\keyword{datasets}
