% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_clusters.R
\name{detect_clusters}
\alias{detect_clusters}
\title{Detecteer kolommen en tekstvakken in PDF-document}
\usage{
detect_clusters(pdf_data, algoritme = "dbscan", ...)
}
\arguments{
\item{pdf_data}{list-object dat het resultaat is van de
\code{\link[pdftools:pdftools]{pdftools::pdf_data()}}-functie}

\item{algoritme}{het algoritme dat gebruikt moet worden om de tekstkolommen
of tekstvakken te detecteren}

\item{...}{algoritme-specifieke argumenten}
}
\value{
Een list-object, waarbij per pagina een Tibble is opgenomen en per
woord is aangegeven tot welk cluster het woord hoort.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Met deze functie detecteer je kolommen
en tekstvakken in een pdf-bestand. Hiervoor moet je eerst het bestand inlezen
met de \code{\link[pdftools:pdftools]{pdftools::pdf_data()}}-functie uit \link{pdftools}.
}
\details{
Op de achtergrond wordt de functie
\link{detect_clusters_page} gebruik
}
\examples{
npo |>
   detect_clusters()
}
