% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_clusters.R
\name{detect_clusters_page}
\alias{detect_clusters_page}
\title{Detecteer kolommen en tekstvakken in PDF-pagina}
\usage{
detect_clusters_page(pdf_data_pagina, algoritme = "dbscan", ...)
}
\arguments{
\item{pdf_data_pagina}{Tibble wat uit elk element van \code{\link[=pdf_data]{pdf_data()}} gehaald
kan worden}

\item{algoritme}{algoritme dat gebruikt moet worden om de tekstkolommen of
tekstvakken te detecteren}

\item{...}{algoritme-specifieke argumenten}
}
\value{
Tibble met per woord aangegeven tot welk cluster het woord hoort.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Met deze functie detecteer je kolommen
en tekstvakken in een pdf-pagina. Hiervoor moet je eerst het bestand inlezen
met de \code{\link[=pdf_data]{pdf_data()}}-functie uit de \link{pdftools}-package.
}
\examples{
npo[[10]] |>
   detect_clusters_page()
}
