% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clusters.R
\name{extract_clusters}
\alias{extract_clusters}
\title{Extract the Text Per Cluster from the \code{\link[=detect_clusters]{detect_clusters()}} object}
\usage{
extract_clusters(pdf_data)
}
\arguments{
\item{pdf_data}{result of the \code{\link[=detect_clusters]{detect_clusters()}}-function or a page of this
result}
}
\value{
If the input is a list of tibbles, a list-object is returned, where
each element is a tibble containing the text extracted from each cluster on
a page. If the input is a single tibble, a tibble is returned directly,
containing the text extracted from each cluster on that page.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
As an end user, you mainly want to work with the text from the different
detected text clusters. This function combines the text from each cluster
into a character string and places it in a Tibble.
}
\examples{
npo[[1]] |>
   detect_clusters() |>
   extract_clusters()
}
