% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clusters_text.R
\name{extract_clusters_text}
\alias{extract_clusters_text}
\title{De tekst per cluster exporteren}
\usage{
extract_clusters_text(pdf_data_pagina_clusters)
}
\arguments{
\item{pdf_data_pagina_clusters}{het resultaat van de functie
\link{detect_clusters_page}}
}
\value{
een Tibble met een aantal records dat gelijk is aan het aantal
gedetecteerde clusters op de pagina
}
\description{
Met deze functie wordt per cluster de tekst aan een character-string
toegevoegd.
}
\examples{
npo[[1]] |>
   detect_clusters_page() |>
   extract_clusters_text()
}
