% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{PdfDocument}
\alias{PdfDocument}
\title{PdfDocument class}
\usage{
PdfDocument(pages = list(), source = NA_character_)
}
\arguments{
\item{pages}{list of \link{PdfPage} objects.}

\item{source}{path or URL the document was read from.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A PDF document: a list of \link{PdfPage} (or, after
\code{\link[=pdf_detect_clusters]{pdf_detect_clusters()}}, \link{PdfClusters}) objects plus the source path or
URL. Created by \code{\link[=pdf_read]{pdf_read()}}.

Pages can be accessed with \code{x[[i]]}, subsets with \code{x[i]}; \code{length(x)}
returns the number of pages.
}
