% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdftoolscolumndetection-package.R
\docType{package}
\name{pdftoolscolumndetection-package}
\alias{pdftoolscolumndetection}
\alias{pdftoolscolumndetection-package}
\title{pdftoolscolumndetection: Het op de juiste manier inlezen van PDF-bestanden waarin mogelijk kolommen of andere tekstvakken voorkomen.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Voor het inlezen van PDF-bestanden bestaat er de package `pdftools`. Bij het inlezen van PDF-bestanden met meerdere kolommen werkt dit package echter niet optimaal. Met de `pdftoolscolumndetection`-package wordt door middel van een Density-Based Spatial Clustering-algoritme bepaald welke woorden (waarschijnlijk) bij elkaar horen.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.ds-ocw.rijkscloud.nl/Datalab/pdftoolscolumndetection}
}

}
\author{
\strong{Maintainer}: Coen Eisma \email{coeneisma@gmail.com} (\href{https://orcid.org/0009-0007-9001-2572}{ORCID})

}
\keyword{internal}
