% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{detect_clusters_page()-object plotten}
\usage{
plot_clusters(pdf_data_pagina_cluster)
}
\arguments{
\item{pdf_data_pagina_cluster}{Een object dat door middel van
detect_clusters_page() is gegenereerd}
}
\value{
een ggplot2-rectangle-plot.
}
\description{
Met deze functie wordt een pagina geplot waarin door middel van de functie
\code{\link[=detect_clusters_page]{detect_clusters_page()}} zijn geclusterd. Elk cluster krijgt een eigen kleur
en cijfer, zodat ze makkelijk visueel te detecten en te vergelijken zijn met
de bron-PDF.
}
\examples{
plot_clusters(npo[[12]])

}
