% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cibap}
\alias{cibap}
\title{CIBAP Report}
\format{
\subsection{\code{cibap}}{

A \link{PdfDocument} with 73 pages. Each page is a \link{PdfPage} whose \code{words}
tibble contains one row per word:
\describe{
\item{width, height}{Width and height of a word}
\item{x, y}{The x and y coordinates of a word. The y-coordinate is measured from the top of the page.}
\item{space}{Indicates whether there is a space after the word. This indicates a line break.}
\item{text}{The word that the metadata refers to.}
\item{font_name, font_size}{The font name and font size of the word.}
}
}
}
\source{
\url{https://www.rijksoverheid.nl/documenten/rapporten/2024/09/16/kwaliteitsagenda-2024-2027-cibap}
}
\usage{
cibap
}
\description{
This dataset is the result of reading the report
\verb{Kwaliteitsagenda 2024-2027 Cibap} with \code{\link[=pdf_read]{pdf_read()}}. The report was
posted on the Rijksoverheid website on September 16, 2024. It was chosen
because it contains several pages with columns and different layouts.
}
\keyword{datasets}
