% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cibap}
\alias{cibap}
\title{CIBAP Report}
\format{
\subsection{\code{cibap}}{

A large list object with 73 elements. Each element contains a tibble with 6 columns and one record for each word on the page.
\describe{
\item{width, height}{Width and height of a word}
\item{x, y}{The x and y coordinates of a word. The x-coordinate is measured from the top.}
\item{space}{Indicates whether there is a space after the word. This indicates a line break.}
\item{text}{The word that the metadata refers to.}
}
}
}
\source{
\url{https://www.rijksoverheid.nl/documenten/rapporten/2024/09/16/kwaliteitsagenda-2024-2027-cibap}
}
\usage{
cibap
}
\description{
This dataset is the result of reading the report
\verb{Kwaliteitsagenda 2024-2027 Cibap}, which was posted on the Rijksoverheid website
on September 16, 2024.
}
\details{
This document was chosen because it contains several pages with
columns and different layouts.
}
\keyword{datasets}
