% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cibap}
\alias{cibap}
\title{CIBAP-rapport}
\format{
\subsection{\code{cibap}}{

Een Large list object met 73 elementen. Elk element bevat een tibble met 6 kolommen en één record per woord dat op de pagina voorkomt.
\describe{
\item{width, height}{Breedte en hoogte van een woord}
\item{x, y}{De x- en y-coördinaten van een woord. De x-coördinaat wordt van boven gemeten.}
\item{space}{Of er na het woord méér dan een spatie aanwezig is.}
\item{text}{Het woord waar de metadata betrekking op heeft.}
}
}
}
\source{
\url{https://www.rijksoverheid.nl/documenten/rapporten/2024/09/16/kwaliteitsagenda-2024-2027-cibap}
}
\usage{
cibap
}
\description{
Deze dataset is het resultaat van het inlezen van van het rapport
\verb{Kwaliteitsagenda 2024-2027 Cibap} die op 16 september 2024 op
de website van de Rijksoverheid is geplaatst.
}
\details{
Dit document is gekozen omdat er verschillende pagina's met
kolommen en afwijkende indelingen zijn.
}
\keyword{datasets}
