% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_clusters.R
\name{detect_clusters_page}
\alias{detect_clusters_page}
\title{Detect Columns and Text Boxes in PDF Page}
\usage{
detect_clusters_page(pdf_data_page, algorithm = "dbscan", ...)
}
\arguments{
\item{pdf_data_page}{list item of the result of the \code{\link[pdftools:pdftools]{pdftools::pdf_data()}}-function}

\item{algorithm}{algorithm to be used to detect text columns or
text boxes}

\item{...}{algorithm-specific arguments}
}
\value{
Tibble indicating which cluster each word belongs to.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} This function detects columns
and text boxes in a PDF page. To do this, you first need to read the file
using the \code{\link[pdftools:pdftools]{pdftools::pdf_data()}}-function from the \link{pdftools} package.
}
\examples{
npo[[10]] |>
   detect_clusters_page()
}
