% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdftextclusteR-package.R
\docType{package}
\name{pdftextclusteR-package}
\alias{pdftextclusteR}
\alias{pdftextclusteR-package}
\title{pdftextclusteR: Properly importing PDF files that may contain columns or other text boxes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

For importing PDF files, the pdftools package is available. However, this package does not work optimally when dealing with PDF files containing multiple columns. The pdftextclusteR package uses a Density-Based Spatial Clustering algorithm to determine which words are (likely) associated with each other.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://coeneisma.github.io/pdftextclusteR/}
}

}
\author{
\strong{Maintainer}: Coen Eisma \email{coeneisma@gmail.com} (\href{https://orcid.org/0009-0007-9001-2572}{ORCID})

}
\keyword{internal}
