% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot the \code{\link[=detect_clusters]{detect_clusters()}} Object}
\usage{
plot_clusters(pdf_data_page_clusters)
}
\arguments{
\item{pdf_data_page_clusters}{a single list-item from the result of
\code{\link[=detect_clusters]{detect_clusters()}}}
}
\value{
a ggplot2 rectangle plot.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function plots a page where the clusters are detected using the
\code{\link[=detect_clusters]{detect_clusters()}} function. Each cluster is assigned a unique color and
number, making them easy to visually detect and compare with the original
PDF.
}
\examples{
npo[[12]] |>
  detect_clusters() |>
  plot_clusters()
}
