% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot the detect_clusters_page() Object}
\usage{
plot_clusters(pdf_data_pagina_cluster)
}
\arguments{
\item{pdf_data_page_cluster}{An object generated by
detect_clusters_page()}
}
\value{
a ggplot2 rectangle plot.
}
\description{
This function plots a page where the clusters are detected using the
\code{\link[=detect_clusters_page]{detect_clusters_page()}} function. Each cluster is assigned a unique color
and number, making them easy to visually detect and compare with
the original PDF.
}
\examples{
plot_clusters(npo[[12]])
}
